Extract the "writer" functionality
